+++ mlir/lib/CAPI/Dialect/Affine.cpp
@@ -8,7 +8,7 @@
 
 #include "mlir-c/Dialect/Affine.h"
 #include "mlir/CAPI/Registration.h"
-#include "mlir/Dialect/Affine/IR/Affine.h"
+#include "mlir/Dialect/Affine/IR/AffineOps.h"
 
 MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Affine, affine,
                                       mlir::affine::AffineDialect)
