+++ mlir/lib/Dialect/Affine/Transforms/AffineScalarReplacement.cpp
@@ -11,7 +11,7 @@
 // redundant loads.
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/Affine/Passes.h"
+#include "mlir/Dialect/Affine/Transforms/Passes.h"
 
 #include "mlir/Analysis/AliasAnalysis.h"
 #include "mlir/Dialect/Affine/Utils.h"
@@ -21,7 +21,7 @@
 namespace mlir {
 namespace affine {
 #define GEN_PASS_DEF_AFFINESCALARREPLACEMENT
-#include "mlir/Dialect/Affine/Passes.h.inc"
+#include "mlir/Dialect/Affine/Transforms/Passes.h.inc"
 } // namespace affine
 } // namespace mlir
