+++ mlir/lib/Dialect/Affine/Transforms/SimplifyAffineMinMax.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/Affine/Passes.h"
+#include "mlir/Dialect/Affine/Transforms/Passes.h"
 
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/Affine/Transforms/Transforms.h"
@@ -227,7 +227,7 @@ struct SimplifyAffineApplyOp : public OpRewritePattern<AffineApplyOp> {
 namespace mlir {
 namespace affine {
 #define GEN_PASS_DEF_SIMPLIFYAFFINEMINMAXPASS
-#include "mlir/Dialect/Affine/Passes.h.inc"
+#include "mlir/Dialect/Affine/Transforms/Passes.h.inc"
 } // namespace affine
 } // namespace mlir
